Move throughput collector persistence into primary instance persistence seam
