--- src/Particular.ThroughputCollector.Persistence.RavenDb/DatabaseConfiguration.cs
@@ -1,6 +1,0 @@
-﻿namespace Particular.ThroughputCollector.Persistence.RavenDb;
-
-public class DatabaseConfiguration(string name)
-{
-    public string Name { get; } = name;
-}
