--- src/Particular.ThroughputCollector.Persistence.RavenDb/Models/ReportConfigurationDocument.cs
@@ -1,6 +1,0 @@
-﻿namespace Particular.ThroughputCollector.Persistence.RavenDb.Models;
-
-class ReportConfigurationDocument
-{
-    public List<string> MaskedStrings { get; set; } = [];
-}
