--- src/Particular.ThroughputCollector.Persistence.RavenDb/Particular.ThroughputCollector.Persistence.RavenDb.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ByteSize" />
-    <!--<PackageReference Include="NServiceBus.CustomChecks" />-->
-    <PackageReference Include="NServiceBus" />
-    <PackageReference Include="RavenDB.Client" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <!-- Private=false & ExcludeAssets=runtime prevent repeatedly including binary dependencies of ThroughputCollector.Persistence in each persister directory -->
-    <ProjectReference Include="..\Particular.ThroughputCollector.Contracts\Particular.ThroughputCollector.Contracts.csproj" />
-    <ProjectReference Include="..\Particular.ThroughputCollector.Persistence\Particular.ThroughputCollector.Persistence.csproj" />
-    <ProjectReference Include="..\ServiceControl.Configuration\ServiceControl.Configuration.csproj"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Artifact Include="$(OutputPath)" DestinationFolder="$(ArtifactsPath)Particular.ServiceControl\ThroughputCollector\Persisters\RavenDb" />
-  </ItemGroup>
-  
-</Project>
