+++ src/ServiceControl.Persistence.RavenDb/Throughput/Models/ReportConfigurationDocument.cs
@@ -1,0 +1,9 @@
+﻿#nullable enable
+namespace ServiceControl.Persistence.RavenDB.Throughput.Models;
+
+using System.Collections.Generic;
+
+class ReportConfigurationDocument
+{
+    public List<string> MaskedStrings { get; set; } = [];
+}
