+++ src/ServiceControl.Persistence.RavenDb/Throughput/ThroughputDatabaseConfiguration.cs
@@ -1,0 +1,7 @@
+﻿#nullable enable
+namespace ServiceControl.Persistence.RavenDB.Throughput;
+
+public class ThroughputDatabaseConfiguration(string name)
+{
+    public string Name { get; } = name;
+}
