+++ src/ServiceControl.Persistence.Tests.RavenDB/ApprovalFiles/RavenPersistedTypes.Verify.approved.txt
@@ -1,3 +1,4 @@
+Particular.ThroughputCollector.Contracts.Endpoint, Particular.ThroughputCollector.Contracts, Version=5.0.0.0, Culture=neutral, PublicKeyToken=null
 ServiceControl.Contracts.CustomChecks.CustomCheck, ServiceControl.Persistence, Version=5.0.0.0, Culture=neutral, PublicKeyToken=null
 ServiceControl.MessageFailures.FailedMessage, ServiceControl.Persistence, Version=5.0.0.0, Culture=neutral, PublicKeyToken=null
 ServiceControl.MessageFailures.GroupComment, ServiceControl.Persistence, Version=5.0.0.0, Culture=neutral, PublicKeyToken=null
