+++ src/ServiceControl.sln
@@ -157,8 +157,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Transports.M
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Persistence", "Particular.ThroughputCollector.Persistence\Particular.ThroughputCollector.Persistence.csproj", "{0EBBFE57-5760-43FC-A646-FE4C49F4BF59}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Persistence.RavenDb", "Particular.ThroughputCollector.Persistence.RavenDb\Particular.ThroughputCollector.Persistence.RavenDb.csproj", "{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.RavenDB", "ServiceControl.RavenDB\ServiceControl.RavenDB.csproj", "{D8610AB6-EB24-4617-8089-F11732D3E033}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Particular.ThroughputCollector", "Particular.ThroughputCollector", "{2D732554-9C3D-4509-8D0B-5595266A4D92}"
@@ -175,13 +173,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollec
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Persistence.InMemory", "Particular.ThroughputCollector.Persistence.InMemory\Particular.ThroughputCollector.Persistence.InMemory.csproj", "{9D134CF7-F1A3-4935-BCBD-458406F8AA97}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Persistence.Tests.RavenDB", "Particular.ThroughputCollector.Persistence.Tests.RavenDB\Particular.ThroughputCollector.Persistence.Tests.RavenDB.csproj", "{F0B1E5C1-E505-44BA-9F01-08531063B586}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Persistence.Tests", "Particular.ThroughputCollector.Persistence.Tests\Particular.ThroughputCollector.Persistence.Tests.csproj", "{A6CAC642-98D9-47B6-85FC-4B51B4A195BF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Api", "ServiceControl.Api\ServiceControl.Api.csproj", "{D101BDEE-85BB-46C5-85E0-6BD8A1B197B1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Particular.ThroughputCollector.Contracts", "Particular.ThroughputCollector.Contracts\Particular.ThroughputCollector.Contracts.csproj", "{6D07AC20-25B3-423F-A88E-64DED435DA06}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Particular.ThroughputCollector.Contracts", "Particular.ThroughputCollector.Contracts\Particular.ThroughputCollector.Contracts.csproj", "{6D07AC20-25B3-423F-A88E-64DED435DA06}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -908,18 +904,6 @@ Global
 		{0EBBFE57-5760-43FC-A646-FE4C49F4BF59}.Release|x64.Build.0 = Release|Any CPU
 		{0EBBFE57-5760-43FC-A646-FE4C49F4BF59}.Release|x86.ActiveCfg = Release|Any CPU
 		{0EBBFE57-5760-43FC-A646-FE4C49F4BF59}.Release|x86.Build.0 = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|x64.Build.0 = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Debug|x86.Build.0 = Debug|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|x64.ActiveCfg = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|x64.Build.0 = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|x86.ActiveCfg = Release|Any CPU
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E}.Release|x86.Build.0 = Release|Any CPU
 		{62AB95DB-2621-4A91-8D84-72D12D9F2D87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{62AB95DB-2621-4A91-8D84-72D12D9F2D87}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{62AB95DB-2621-4A91-8D84-72D12D9F2D87}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -956,18 +940,6 @@ Global
 		{9D134CF7-F1A3-4935-BCBD-458406F8AA97}.Release|x64.Build.0 = Release|Any CPU
 		{9D134CF7-F1A3-4935-BCBD-458406F8AA97}.Release|x86.ActiveCfg = Release|Any CPU
 		{9D134CF7-F1A3-4935-BCBD-458406F8AA97}.Release|x86.Build.0 = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|x64.Build.0 = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Debug|x86.Build.0 = Debug|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|x64.ActiveCfg = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|x64.Build.0 = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|x86.ActiveCfg = Release|Any CPU
-		{F0B1E5C1-E505-44BA-9F01-08531063B586}.Release|x86.Build.0 = Release|Any CPU
 		{A6CAC642-98D9-47B6-85FC-4B51B4A195BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A6CAC642-98D9-47B6-85FC-4B51B4A195BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6CAC642-98D9-47B6-85FC-4B51B4A195BF}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1079,13 +1051,11 @@ Global
 		{9682F7C5-C631-4924-A698-4C355BB14540} = {9AF9D3C7-E859-451B-BA4D-B954D289213A}
 		{D45F51F6-10E5-4976-9E7D-C15A9012B84F} = {C22BFC89-851E-4D42-A878-C13B3F206908}
 		{0EBBFE57-5760-43FC-A646-FE4C49F4BF59} = {C22BFC89-851E-4D42-A878-C13B3F206908}
-		{CD5CCAD6-4ED4-4555-803E-3E4C9DBB8A3E} = {C22BFC89-851E-4D42-A878-C13B3F206908}
 		{C22BFC89-851E-4D42-A878-C13B3F206908} = {2D732554-9C3D-4509-8D0B-5595266A4D92}
 		{94BDFE10-0672-4299-A7D8-2283789F4549} = {2D732554-9C3D-4509-8D0B-5595266A4D92}
 		{62AB95DB-2621-4A91-8D84-72D12D9F2D87} = {2D732554-9C3D-4509-8D0B-5595266A4D92}
 		{E64B0BEF-B729-411D-9FE9-7AB52FE7E6C6} = {94BDFE10-0672-4299-A7D8-2283789F4549}
 		{9D134CF7-F1A3-4935-BCBD-458406F8AA97} = {C22BFC89-851E-4D42-A878-C13B3F206908}
-		{F0B1E5C1-E505-44BA-9F01-08531063B586} = {94BDFE10-0672-4299-A7D8-2283789F4549}
 		{A6CAC642-98D9-47B6-85FC-4B51B4A195BF} = {94BDFE10-0672-4299-A7D8-2283789F4549}
 		{D101BDEE-85BB-46C5-85E0-6BD8A1B197B1} = {9AF9D3C7-E859-451B-BA4D-B954D289213A}
 		{6D07AC20-25B3-423F-A88E-64DED435DA06} = {2D732554-9C3D-4509-8D0B-5595266A4D92}
